+++ rust-competitive-helper-util/src/lib.rs
@@ -90,21 +90,20 @@ pub struct Task {
     pub languages: Languages,
 }
 
-pub fn read_from_file<P: AsRef<Path>>(filename: P) -> String {
-    fs::read_to_string(&filename)
-        .unwrap_or_else(|_| panic!("Can't read file: {:?}", filename.as_ref().as_os_str()))
+pub fn read_from_file<P: AsRef<Path>>(filename: P) -> Option<String> {
+    fs::read_to_string(&filename).ok()
 }
 
-pub fn read_lines<P>(filename: P) -> Vec<String>
+pub fn read_lines<P>(filename: P) -> Result<Vec<String>, String>
 where
     P: AsRef<Path> + Display,
 {
-    let file = File::open(&filename).unwrap_or_else(|_| panic!("Can't read file: '{}'", filename));
+    let file = File::open(&filename).map_err(|_| format!("Can't read file: '{}'", filename))?;
     let mut res = Vec::new();
     for line in io::BufReader::new(file).lines() {
         res.push(line.unwrap());
     }
-    res
+    Ok(res)
 }
 
 pub fn all_rs_files_in_dir<P>(path: P) -> Vec<String>