+++ src/archiver.rs
@@ -6,7 +6,9 @@ use itertools::Itertools;
 use rust_competitive_helper_util::Task;
 use std::collections::BTreeMap;
 use std::fs;
+use std::fs::{read_dir, remove_dir_all, rename};
 use std::io::{BufRead, BufReader};
+use std::iter::once;
 
 fn contest_name(group: &str) -> String {
     match group.find('-') {
@@ -16,15 +18,18 @@ fn contest_name(group: &str) -> String {
 }
 
 fn contest_list() -> Vec<(String, Vec<String>)> {
-    let lines = rust_competitive_helper_util::read_lines("Cargo.toml");
     let mut result = BTreeMap::new();
-    for line in lines {
-        if !line.starts_with("    ") {
-            continue;
-        }
-        let line = line.trim().as_bytes();
-        let task_name = String::from_utf8_lossy(&line[1..line.len() - 2]).to_string();
-        let main = fs::File::open(format!("{}/src/main.rs", task_name));
+    for task_name in read_dir("tasks").unwrap().map(|entry| {
+        entry
+            .unwrap()
+            .path()
+            .file_stem()
+            .unwrap()
+            .to_str()
+            .unwrap()
+            .to_string()
+    }) {
+        let main = fs::File::open(format!("tasks/{}/src/main.rs", task_name));
         if main.is_err() {
             continue;
         }
@@ -53,32 +58,26 @@ fn contest_list() -> Vec<(String, Vec<String>)> {
 const OPTIONS: [&str; 4] = ["Skip", "Delete", "Archive only", "Archive and tests"];
 
 fn find_additional_solution_files(task_name: &str) -> Vec<String> {
-    rust_competitive_helper_util::all_rs_files_in_dir(format!("{}/src", task_name))
+    rust_competitive_helper_util::all_rs_files_in_dir(format!("tasks/{}/src", task_name))
         .into_iter()
         .filter(|file| file != "main.rs" && file != "tester.rs")
         .collect()
 }
 
-fn ask_archive(task_name: String) {
-    let selection = Select::with_theme(&ColorfulTheme::default())
-        .with_prompt(format!("Task {}", task_name))
-        .default(2)
-        .items(&OPTIONS[..])
-        .interact_on(&Term::stdout())
-        .unwrap();
-
+fn ask_archive(task_name: String, selection: usize) {
     if selection == 0 {
         return;
     }
     if selection >= 2 {
         let now = Utc::now();
         let mut main =
-            rust_competitive_helper_util::read_lines(format!("{}/src/main.rs", task_name));
+            rust_competitive_helper_util::read_lines(format!("tasks/{}/src/main.rs", task_name))
+                .unwrap();
         let task =
             serde_json::from_str::<Task>(main[0].chars().skip(2).collect::<String>().as_str())
                 .unwrap();
         let path = format!(
-            "archive/{}/{}/{}.{}.{} - {}",
+            "archive/{}/{:02}/{:02}.{:02}.{} - {}",
             now.year(),
             now.month(),
             now.day(),
@@ -93,13 +92,19 @@ fn ask_archive(task_name: String) {
             main.clone(),
         );
         for file in find_additional_solution_files(&task_name) {
-            let content =
-                rust_competitive_helper_util::read_lines(format!("{}/src/{}", task_name, file));
+            let content = rust_competitive_helper_util::read_lines(format!(
+                "tasks/{}/src/{}",
+                task_name, file
+            ))
+            .unwrap();
             rust_competitive_helper_util::write_lines(format!("{}/{}.rs", path, file), content);
         }
         if selection == 3 {
-            let tester =
-                rust_competitive_helper_util::read_lines(format!("{}/src/tester.rs", task_name));
+            let tester = rust_competitive_helper_util::read_lines(format!(
+                "tasks/{}/src/tester.rs",
+                task_name
+            ))
+            .unwrap();
             main.push("mod tester {".to_string());
             main.extend_from_slice(tester.as_slice());
             main.push("}".to_string());
@@ -133,15 +138,16 @@ fn ask_archive(task_name: String) {
                 format!("algo_lib/tests/{}.rs", task_name),
                 test_lines,
             );
-            let from = format!("{}/tests", task_name);
-            std::fs::rename(from, format!("algo_lib/tests/{}", task_name)).unwrap();
+            let from = format!("tasks/{}/tests", task_name);
+            rename(from, format!("algo_lib/tests/{}", task_name)).unwrap();
         }
     }
-    std::fs::remove_dir_all(format!("{}/", task_name)).unwrap();
+    remove_dir_all(format!("tasks/{}/", task_name)).unwrap();
 
     let lines = rust_competitive_helper_util::read_lines("Cargo.toml")
+        .unwrap()
         .into_iter()
-        .filter(|line| line != &format!("    \"{}\",", task_name))
+        .filter(|line| line != &format!("    \"tasks/{}\",", task_name))
         .collect_vec();
     rust_competitive_helper_util::write_lines("Cargo.toml", lines);
 }
@@ -171,7 +177,53 @@ pub fn archive() {
     let mut tasks = contest_list[selection].1.clone();
     tasks.sort();
 
-    for task in tasks {
-        ask_archive(task);
+    let mut selection = vec![2; tasks.len()];
+
+    let mut last = 0;
+
+    loop {
+        let id = Select::with_theme(&ColorfulTheme::default())
+            .with_prompt("Change:")
+            .default(last)
+            .items(
+                once("Done".to_string())
+                    .chain(
+                        once("Default".to_string()).chain(
+                            tasks
+                                .iter()
+                                .enumerate()
+                                .map(|(i, s)| format!("{} ({})", s, OPTIONS[selection[i]])),
+                        ),
+                    )
+                    .collect::<Vec<_>>()
+                    .as_slice(),
+            )
+            .interact_on(&Term::stdout())
+            .unwrap();
+        let option = match id {
+            0 => break,
+            _ => Select::with_theme(&ColorfulTheme::default())
+                .with_prompt("Select option:")
+                .default(if id > 1 {
+                    selection[id - 2]
+                } else if selection.iter().all(|&x| x == selection[0]) {
+                    selection[0]
+                } else {
+                    2
+                })
+                .items(&OPTIONS[..])
+                .interact_on(&Term::stdout())
+                .unwrap(),
+        };
+        if id == 1 {
+            selection.fill(option);
+        } else {
+            selection[id - 2] = option;
+        }
+        last = id;
+    }
+
+    for (task, opt) in tasks.into_iter().zip(selection) {
+        ask_archive(task, opt);
     }
 }